+++ app/lib/shared/versions.dart
@@ -24,7 +24,7 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const _acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2024.11.26',
+  '2024.12.04',
   // Fallback runtime versions.
   '2024.11.21',
   '2024.11.18',
@@ -62,8 +62,8 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '3.5.0';
-final String toolStableDartSdkVersion = '3.5.4';
-final String toolStableFlutterSdkVersion = '3.24.5';
+final String toolStableDartSdkVersion = '3.6.0-334.4.beta';
+final String toolStableFlutterSdkVersion = '3.27.0-0.2.pre';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);
@@ -74,4 +74,4 @@ final semanticToolStableFlutterSdkVersion =
 final String panaVersion = pana.packageVersion;
 
 // keep in-sync with pkg/pub-worker/lib/src/bin/pana_wrapper.dart
-final String dartdocVersion = '8.1.0';
+final String dartdocVersion = '8.3.0';